+++ piccolo/apps/migrations/commands/new.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import datetime
 import os
+import string
 import sys
 import typing as t
 from dataclasses import dataclass
@@ -31,8 +32,9 @@ JINJA_ENV = jinja2.Environment(
     loader=jinja2.FileSystemLoader(searchpath=TEMPLATE_DIRECTORY),
 )
 
-
 MIGRATION_MODULES: t.Dict[str, ModuleType] = {}
+
+VALID_PYTHON_MODULE_CHARACTERS = string.ascii_lowercase + string.digits + "_"
 
 
 def render_template(**kwargs):
@@ -60,6 +62,13 @@ class NewMigrationMeta:
     migration_path: str
 
 
+def now():
+    """
+    In a separate function so it's easier to patch in tests.
+    """
+    return datetime.datetime.now()
+
+
 def _generate_migration_meta(app_config: AppConfig) -> NewMigrationMeta:
     """
     Generates the migration ID and filename.
@@ -68,13 +77,25 @@ def _generate_migration_meta(app_config: AppConfig) -> NewMigrationMeta:
     # chance that the IDs would clash if the migrations are generated
     # programatically in quick succession (e.g. in a unit test), so they had
     # to be added. The trade off is a longer ID.
-    _id = datetime.datetime.now().strftime("%Y-%m-%dT%H:%M:%S:%f")
+    _id = now().strftime("%Y-%m-%dT%H:%M:%S:%f")
 
     # Originally we just used the _id as the filename, but colons aren't
     # supported in Windows, so we need to sanitize it. We don't want to
     # change the _id format though, as it would break existing migrations.
     # The filename doesn't have any special significance - only the id matters.
-    filename = _id.replace(":", "-")
+    cleaned_id = _id.replace(":", "_").replace("-", "_").lower()
+
+    # Just in case the app name contains characters which aren't valid for
+    # a Python module.
+    cleaned_app_name = "".join(
+        [
+            i
+            for i in app_config.app_name.lower().replace("-", "_")
+            if i in VALID_PYTHON_MODULE_CHARACTERS
+        ]
+    )
+
+    filename = f"{cleaned_app_name}_{cleaned_id}"
 
     path = os.path.join(app_config.migrations_folder_path, f"{filename}.py")