+++ piccolo/query/__init__.py
@@ -18,3 +18,26 @@ from .methods import (
     TableExists,
     Update,
 )
+from .mixins import OrderByRaw
+
+__all__ = [
+    "Alter",
+    "Avg",
+    "Count",
+    "Create",
+    "CreateIndex",
+    "Delete",
+    "DropIndex",
+    "Exists",
+    "Insert",
+    "Max",
+    "Min",
+    "Objects",
+    "OrderByRaw",
+    "Query",
+    "Raw",
+    "Select",
+    "Sum",
+    "TableExists",
+    "Update",
+]