+++ tests/utils/test_pydantic.py
@@ -2,6 +2,7 @@ import decimal
 from unittest import TestCase
 
 import pydantic
+import pytest
 from pydantic import ValidationError
 
 from piccolo.columns import (
@@ -264,6 +265,17 @@ class TestJSONColumn(TestCase):
             pydantic_model.schema()["properties"]["features"]["format"],
             "json",
         )
+
+    def test_null_value(self):
+        class Movie(Table):
+            meta = JSON(null=True)
+            meta_b = JSONB(null=True)
+
+        pydantic_model = create_pydantic_model(table=Movie)
+        movie = pydantic_model(meta=None, meta_b=None)
+
+        self.assertIsNone(movie.meta)
+        self.assertIsNone(movie.meta_b)
 
 
 class TestExcludeColumns(TestCase):
@@ -690,3 +702,37 @@ class TestSchemaExtraKwargs(TestCase):
 
         model = create_pydantic_model(Band, visible_columns=("name",))
         self.assertEqual(model.schema()["visible_columns"], ("name",))
+
+
+class TestPydanticExtraFields(TestCase):
+    def test_pydantic_extra_fields(self):
+        """
+        Make sure that the value of ``extra`` in the config class
+        is correctly propagated to the generated model.
+        """
+
+        class Band(Table):
+            name = Varchar()
+
+        for v in ["ignore", "allow", "forbid"]:
+
+            class MyConfig(pydantic.BaseConfig):
+                extra = v
+
+            model = create_pydantic_model(Band, pydantic_config_class=MyConfig)
+            self.assertEqual(model.Config.extra, v)
+
+    def test_pydantic_invalid_extra_fields(self):
+        """
+        Make sure that invalid values for ``extra`` in the config class
+        are rejected.
+        """
+
+        class Band(Table):
+            name = Varchar()
+
+        class MyConfig(pydantic.BaseConfig):
+            extra = "foobar"
+
+        with pytest.raises(ValueError):
+            create_pydantic_model(Band, pydantic_config_class=MyConfig)