+++ tests/utils/test_sql_values.py
@@ -1,7 +1,10 @@
+import time
 from enum import Enum
 from unittest import TestCase
 
-from piccolo.columns.column_types import JSON, JSONB, Integer, Varchar
+import pytest
+
+from piccolo.columns.column_types import JSON, JSONB, Array, Integer, Varchar
 from piccolo.table import Table
 from piccolo.utils.sql_values import convert_to_sql_value
 
@@ -48,10 +51,10 @@ class TestConvertToSQLValue(TestCase):
         """
 
         class Colour(Enum):
-            red = "red"
+            red = "r"
 
         self.assertEqual(
-            convert_to_sql_value(value=Colour.red, column=Varchar()), "red"
+            convert_to_sql_value(value=Colour.red, column=Varchar()), "r"
         )
 
     def test_other(self):
@@ -62,3 +65,37 @@ class TestConvertToSQLValue(TestCase):
             convert_to_sql_value(value=1, column=Integer()),
             1,
         )
+
+    def test_convert_enum_list(self):
+        """
+        It's possible to have a list of enums when using ``Array`` columns.
+        """
+
+        class Colour(Enum):
+            red = "r"
+            green = "g"
+            blue = "b"
+
+        self.assertEqual(
+            convert_to_sql_value(
+                value=[Colour.red, Colour.green, Colour.blue],
+                column=Array(Varchar()),
+            ),
+            ["r", "g", "b"],
+        )
+
+    @pytest.mark.speed
+    def test_convert_large_list(self):
+        """
+        Large lists are problematic. We need to check each value in the list,
+        but as efficiently as possible.
+        """
+        start = time.time()
+
+        convert_to_sql_value(
+            value=[i for i in range(1000)],
+            column=Array(Varchar()),
+        )
+
+        duration = time.time() - start
+        print(duration)